Extract program info from utils/tx.ts to remove web3js dependency
